requirements.txt
numpy<2
pyyaml
tqdm
pillow
opencv-python
scikit-image
tensorboard
matplotlib
jupyter
